Add docs `Week::class`.
