+++ tests/Input/Week/AttributeTest.php
@@ -153,26 +153,6 @@ final class AttributeTest extends TestCase
         );
     }
 
-    public function testMax(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" max="value">
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->max('value')->render()
-        );
-    }
-
-    public function testMin(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" min="value">
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->min('value')->render()
-        );
-    }
-
     public function testName(): void
     {
         Assert::equalsWithoutLE(
@@ -190,26 +170,6 @@ final class AttributeTest extends TestCase
             <input id="week-6582f2d099e8b" type="week" readonly>
             HTML,
             Week::widget()->id('week-6582f2d099e8b')->readonly()->render()
-        );
-    }
-
-    public function testRender(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week">
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testRequired(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="week-6582f2d099e8b" type="week" required>
-            HTML,
-            Week::widget()->id('week-6582f2d099e8b')->required()->render()
         );
     }
